Add partial classes for netstandard21 target
