+++ src/libraries/Microsoft.Extensions.FileProviders.Physical/src/PhysicalFilesWatcher.netstandard21.cs
@@ -1,0 +1,35 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Threading.Tasks;
+using Microsoft.Extensions.Primitives;
+
+namespace Microsoft.Extensions.FileProviders
+{
+    /// <summary>
+    ///     <para>
+    ///     A file watcher that watches a physical filesystem for changes.
+    ///     </para>
+    ///     <para>
+    ///     Triggers events on <see cref="IChangeToken" /> when files are created, change, renamed, or deleted.
+    ///     </para>
+    /// </summary>
+    public partial class PhysicalFilesWatcher : IAsyncDisposable
+    {
+        public async ValueTask DisposeAsync()
+        {
+            await DisposeAsyncCore();
+            Dispose(false);
+            GC.SuppressFinalize(this);
+        }
+
+        protected virtual async ValueTask DisposeAsyncCore()
+        {
+            await _timer.DisposeAsync().ConfigureAwait(false);
+            await _fileWatcher.DisposeAsync().ConfigureAwait(false);
+        }
+    }
+}
